made clustering section compatible with radial pos
